30% Group Project/Group A1/WebApp/requirements.txt
Flask==3.0.0
mysql-connector-python==8.0.33
requests==2.31.0
